pyproject.toml
[build-system]
requires = ["hatchling","hatch-build-scripts"]
build-backend = "hatchling.build"

[project]
name = "moonshot"
version = "0.1.0"
authors = [
    { name="The Moonshot Team", email="our.moonshot.team@gmail.com" }
]
description = "A simple and modular tool to evaluate and red-team any LLM application."
readme = "README.md"
license = { text = "Apache-2.0" }
requires-python = ">=3.11"
classifiers = [
    "Development Status :: 3 - Alpha",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "License :: OSI Approved :: Apache Software License"
]
dependencies = [
    "openai==0.28.1",
    "python-slugify==8.0.4",
    "bert-score==0.3.13",
    "nltk==3.8.1",
    "py-readability-metrics==1.4.5",
    "rouge==1.0.1",
    "rouge-score==0.1.2",
    "aiohttp==3.9.1",
    "aiometer==0.4.0",
    "python-dotenv==1.0.0",
    "anthropic==0.7.8",
    "jinja2==3.1.3",
    "spacy==3.7.2",
    "numpy==1.26.3",
    "ijson==3.2.3",
]

[project.optional-dependencies]
web-api = [
    "typing-extensions==4.9.0",
    "fastapi==0.108.0",
    "uvicorn==0.25.0",
    "dependency-injector==4.41.0",
]
cli = [
    "cmd2==2.4.3",
    "rich==13.7.0",
]
all = [
    "moonshot[web-api]",
    "moonshot[cli]",
]

[project.urls]
Repository = "https://github.com/moonshot-admin/moonshot"
Documentation = "https://moonshot-admin.github.io/moonshot/"
Issues = "https://github.com/moonshot-admin/moonshot/issues"

[tool.hatch.metadata]
allow-direct-references = true